add all the alpaca apys
