+++ src/data/alpacaPools.json
@@ -1,0 +1,26 @@
+[
+  {
+    "name": "sALPACA",
+    "token": "0x6F695Bd5FFD25149176629f8491A5099426Ce7a7",
+    "oracle": "tokens",
+    "oracleId": "ALPACA",
+    "decimals": "1e18",
+    "poolId": 5
+  },
+  {
+    "name": "ibBNB",
+    "token": "0xd7D069493685A581d27824Fc46EdA46B7EfC0063",
+    "oracle": "tokens",
+    "oracleId": "WBNB",
+    "decimals": "1e18",
+    "poolId": 1
+  },
+  {
+    "name": "ibBUSD",
+    "token": "0x7C9e73d4C71dae564d41F78d56439bB4ba87592f",
+    "oracle": "tokens",
+    "oracleId": "BUSD",
+    "decimals": "1e18",
+    "poolId": 3
+  }
+]
